added start stop linking implementation to drone session manager.
